Refactoring. Get rid of redundant information.
Services now have an identifier and a display name. DesktopStarters should only have a display name.

The SetupBuilder itself only needs the basic identifier and application name. Both can be derived. But it also has a setupName to automatically add the version number to the identifier if no other name is configured.