Rafactor: make version available as task variable
The refactoring allows to define the version variable in a platform specific task. There is now the choice to set the artifact version specific to a platform, for the whole SetupBuilder and therefore each platform or globally in the project.

The change allows a more fine grained setup where version numbers are in a special format (msi).