Use SFAuthentication instead of AppleScript Helper Tool
Modified the Preferences Pane Authentication Method. The reason for this huge change - to a deprecated API - is that the AppleScript variant will not get notarized due to the binary being created with an 10.9 SDK.

SMJob can not be used due to being run in the preferences panel which disallows it. The “old” method is still recommended. It uses a helper file that has privilige elevation.